movied queriedSongs to separate file, functioning
