+++ queries/songs.js
@@ -11,49 +11,49 @@ async function getAllSongs() {
     }
 }
 
-//  BONUS ASC/DESC QUERIES
-async function getQueriedSongs(obj) {
-    if(obj.order){
-        if(obj.order === "asc"){
-            try{
-                const ascOrder = await database.any('SELECT * FROM songs ORDER BY name ASC') 
-                return ascOrder
-            } catch(err){
-                return err
-            }
-        }
-        else if(obj.order === "desc"){
-            try{
-                const descOrder = await database.any('SELECT * FROM songs ORDER BY name DESC') 
-                return descOrder
-            } catch(err){
-                return err
-            }
+// //  BONUS ASC/DESC QUERIES
+// async function getQueriedSongs(obj) {
+//     if(obj.order){
+//         if(obj.order === "asc"){
+//             try{
+//                 const ascOrder = await database.any('SELECT * FROM songs ORDER BY name ASC') 
+//                 return ascOrder
+//             } catch(err){
+//                 return err
+//             }
+//         }
+//         else if(obj.order === "desc"){
+//             try{
+//                 const descOrder = await database.any('SELECT * FROM songs ORDER BY name DESC') 
+//                 return descOrder
+//             } catch(err){
+//                 return err
+//             }
 
-        }
-    }
-    if(obj.is_favorite){
-        if(obj.is_favorite === "true"){
-          try{
-            const isFavorite = await database.any('SELECT * FROM songs WHERE is_favorite = true')
-            return isFavorite
-          } catch(err){
-            return err
-          }
-        }
-        else if(obj.is_favorite === "false"){
-            try{
-              const isNotFavorite = await database.any('SELECT * FROM songs WHERE is_favorite = false')
-              return isNotFavorite
-            } catch(err){
-              return err
-            }
-          }
-    }
-    else{
-        return undefined
-    }
-}
+//         }
+//     }
+//     if(obj.is_favorite){
+//         if(obj.is_favorite === "true"){
+//           try{
+//             const isFavorite = await database.any('SELECT * FROM songs WHERE is_favorite = true')
+//             return isFavorite
+//           } catch(err){
+//             return err
+//           }
+//         }
+//         else if(obj.is_favorite === "false"){
+//             try{
+//               const isNotFavorite = await database.any('SELECT * FROM songs WHERE is_favorite = false')
+//               return isNotFavorite
+//             } catch(err){
+//               return err
+//             }
+//           }
+//     }
+//     else{
+//         return undefined
+//     }
+// }
 
 // Get one specific song based on id -> database.one() -> $1 (for substituting variables)
 async function getOneSong(idValue) {
@@ -130,13 +130,53 @@ async function deleteSong(idValue) {
 
 // UPDATE/ EDIT A SONG (idValue, updatedObj)-> .one() -> use UPDATE, SET, WHERE and RETURNING keywords
 async function updateSong(idValue, updatedObj) {
-    try {
-        const updatedSong = database.one('UPDATE songs SET name = $2, artist = $3, album = $4, time = $5, is_favorite = $6 WHERE id = $1 RETURNING *', [idValue, updatedObj.name, updatedObj.artist, updatedObj.album, updatedObj.time, updatedObj.is_favorite])
-
-        return updatedSong
-    } catch(err) {
-        return err
+    if(!updatedObj.album && !updatedObj.time){
+        try {
+            const newSong = await database.one(
+            'INSERT INTO songs(name, artist, is_favorite) VALUES ($1, $2, $3) RETURNING *', 
+            [updatedObj.name, updatedObj.artist, updatedObj.is_favorite]
+            )
+            return newSong
+        } catch(err) {
+            return err
+        }
     }
+    else if(!updatedObj.album || !updatedObj.time){
+        if(!updatedObj.album){
+            try {
+                const newSong = await database.one(
+                'INSERT INTO songs(name, artist, time, is_favorite) VALUES ($1, $2, $3, $4) RETURNING *', 
+                [updatedObj.name, updatedObj.artist, updatedObj.time, updatedObj.is_favorite]
+                )
+                return newSong
+            } catch(err) {
+                return err
+            }
+        }
+        else if(!updatedObj.time){
+            try {
+                const newSong = await database.one(
+                'INSERT INTO songs(name, artist, album, is_favorite) VALUES ($1, $2, $3, $4) RETURNING *', 
+                [updatedObj.name, updatedObj.artist, updatedObj.album, updatedObj.is_favorite]
+                )
+                return newSong
+            } catch(err) {
+                return err
+            }
+        }
+    }
+    else {
+        try {
+            const updatedSong = database.one('UPDATE songs SET name = $2, artist = $3, album = $4, time = $5, is_favorite = $6 WHERE id = $1 RETURNING *', [idValue, updatedObj.name, updatedObj.artist, updatedObj.album, updatedObj.time, updatedObj.is_favorite])
+    
+            return updatedSong
+        } catch(err) {
+            return err
+        }
+
+    }
+
+   
 }
 
 
@@ -147,5 +187,4 @@ module.exports = {
     createSong,
     deleteSong,
     updateSong,
-    getQueriedSongs,
 }
